backend/requirements.txt
Django==4.2
djangorestframework==3.14
celery==5.3
redis==4.5
psycopg2-binary==2.9
requests==2.31
python-dotenv==1.0
Jinja2==3.1
